add Teaser (grades)
A Teaser could be a standalone code. This seems like a good idea, because if the sentence only teasers information it is hard to tell what the purpose of the information is, it's only possible to guess most of the times.
But it is relevant to know of the existence of teasers, since the more teasers and the less conclusions, the likelier it is that the authors have conclusions written in some other sections. But not the conclusion section. Meaning the mental model would seem to be to use the conclusion section like a summary or a big teaser to lure the reader into reading the paper maybe?